fix build and runtime errors

+++ collector/insight/process.go
@@ -165,7 +165,7 @@ func getProcStartTime(proc *process.Process) (float64, error) {
 	}
 	fields := strings.Fields(string(contents))
 	if startTime, err := strconv.ParseFloat(fields[21], 64); err == nil {
-		return startTime / process.ClockTicks, err
+		return startTime / float64(process.ClockTicks), err
 	}
 	return 0, err
 }
